Added Event Pagination and Forms

+++ Application/Classes/Alcohol.cs
@@ -11,7 +11,7 @@ namespace Classes
         public int Percentage { get; private set; }
         public int Age { get; private set; }
         
-        public Alcohol(int id, byte[] picture, string name, int size, double price, int percentage, int age) : base(id, picture, name, size, price)
+        public Alcohol(int id, byte[] picture, string name, int size, decimal price, int percentage, int age) : base(id, picture, name, size, price)
         {
             Percentage = percentage;
             Age = age;
